feat(core): Reduce number of subfolders
