--- src/nwp_consumer/internal/core/__init__.py
@@ -1,1 +1,0 @@
-"""Core of the package."""
