--- src/nwp_consumer/internal/core/domain/__init__.py
@@ -1,10 +1,0 @@
-"""Struct definitions for core domain objects.
-
-Not every struct is a domain model! Only those involved in the business logic.
-"""
-
-from .repometadata import *
-from .storemetadata import *
-from .postprocess import *
-from .notification import *
-from .parameters import *
