--- src/nwp_consumer/internal/core/service/__init__.py
@@ -1,5 +1,0 @@
-"""Interfaces for core service implementations.
-
-The services module defines abstract interfaces that specify the signatures
-any service implementations must obey in order to interact with the core.
-"""
