+++ src/nwp_consumer/internal/repositories/notification/stdout.py
@@ -4,8 +4,8 @@ import logging
 
 from returns.io import IOResultE, IOSuccess
 
-from nwp_consumer.internal.core import domain
-from nwp_consumer.internal.core.ports import NotificationRepository
+from nwp_consumer.internal import entities
+from nwp_consumer.internal.ports import NotificationRepository
 
 log = logging.getLogger(__package__)
 
@@ -15,7 +15,7 @@ class StdoutNotificationRepository(NotificationRepository):
 
     def notify(
             self,
-            message: domain.StoreCreatedNotification | domain.StoreAppendedNotification,
+            message: entities.StoreCreatedNotification | entities.StoreAppendedNotification,
     ) -> IOResultE[str]:
         """Overrides the corresponding method in the parent class."""
         log.info(f"Sent notification (type: {message.__class__.__name__}): {message}")
