+++ src/nwp_consumer/internal/service/__init__.py
@@ -1,3 +1,5 @@
-__all__ = ["NWPConsumerService"]
+"""Interfaces for core service implementations.
 
-from .consumer import NWPConsumerService
+The services module defines abstract interfaces that specify the signatures
+any service implementations must obey in order to interact with the core.
+"""
