+++ src/nwp_consumer/internal/service/memory.py
